# etc/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Homer-App User interface for WEB AI",
    "title": "Homer-App",
    "license": {
      "name": "AGPL",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "version": "1.1.2"
  },
  "host": "localhost:9080",
  "basePath": "/api/v3",
  "paths": {
...
    "/alias": {
      "get": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Get all aliases",
        "produces": [
          "application/json"
        ],
        "tags": [
          "alias"
        ],
        "operationId": "aliasGetAllAlias",
        "responses": {
          "200": {
            "description": "AliasStructList",
            "schema": {
              "$ref": "#/definitions/AliasStructList"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Adds alias to system",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alias"
        ],
        "operationId": "aliasAddAlias",
        "parameters": [
          {
            "type": "object",
            "description": "AliasStruct parameters",
            "name": "AliasStruct",
            "in": "body",
            "required": true,
            "schema": {
              "description": "AliasStruct parameters",
              "type": "object",
              "$ref": "#/definitions/AliasStruct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "AliasCreationSuccessResponse",
            "schema": {
              "$ref": "#/definitions/AliasCreationSuccessResponse"
            }
          },
          "401": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
    "/alias/{guid}": {
      "put": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "Update an existing alias based upon the lookup",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alias"
        ],
        "operationId": "aliasUpdateAlias",
        "parameters": [
          {
            "type": "string",
            "description": "uuid of the Alias to update",
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "description": "uuid of the Alias to update",
              "type": "string"
            }
          },
          {
            "type": "object",
            "description": "AliasStruct parameters",
            "name": "AliasStruct",
            "in": "body",
            "required": true,
            "schema": {
              "description": "AliasStruct parameters",
              "type": "object",
              "$ref": "#/definitions/AliasStruct"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "AliasUpdateSuccessResponse",
            "schema": {
              "$ref": "#/definitions/AliasUpdateSuccessResponse"
            }
          },
          "401": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "bearer": [
              "[]"
            ]
          },
          {
            "SecurityDefinitions": []
          },
          {
            "bearer": []
          },
          {
            "type": [
              "apiKey"
            ]
          },
          {
            "name": [
              "Authorization"
            ]
          },
          {
            "in": [
              "header"
            ]
          }
        ],
        "description": "delete an alias based upon guid",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "alias"
        ],
        "operationId": "aliasDeleteAlias",
        "parameters": [
          {
            "type": "string",
            "description": "guid of the alias to delete",
            "name": "guid",
            "in": "path",
            "required": true,
            "schema": {
              "description": "guid of the alias to delete",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "AliasDeleteSuccessResponse",
            "schema": {
              "$ref": "#/definitions/AliasDeleteSuccessResponse"
            }
          },
          "401": {
            "description": "FailureResponse",
            "schema": {
              "$ref": "#/definitions/FailureResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AliasCreationSuccessResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f2d0a540-bf21-4c0d-ac73-8696ea10855a"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully created alias"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "AliasDeleteSuccessResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f2d0a540-bf21-4c0d-ac73-8696ea10855a"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully deleted alias"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "AliasStruct": {
      "type": "object",
      "required": [
        "alias",
        "ip",
        "port",
        "mask",
        "captureID"
      ],
      "properties": {
        "alias": {
          "description": "alias name",
          "type": "string",
          "x-go-name": "Alias",
          "example": "verizon"
        },
        "captureID": {
          "type": "string",
          "x-go-name": "CaptureID",
          "example": "0"
        },
        "guid": {
          "type": "string",
          "x-go-name": "GUID"
        },
        "id": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Id"
        },
        "ip": {
          "type": "string",
          "x-go-name": "IP",
          "example": "192.168.10.20"
        },
        "mask": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Mask",
          "example": 32
        },
        "port": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Port",
          "example": 5060
        },
        "status": {
          "type": "boolean",
          "x-go-name": "Status"
        }
      },
      "x-go-name": "TableAlias",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "AliasStructList": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/AliasStruct"
          },
          "x-go-name": "Data"
        }
      },
      "x-go-name": "TableAliasList",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
    "AliasUpdateSuccessResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "x-go-name": "Data",
          "example": "f2d0a540-bf21-4c0d-ac73-8696ea10855a"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message",
          "example": "successfully updated alias"
        }
      },
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
    "FailureResponse": {
      "type": "object",
      "properties": {
        "error": {
          "description": "errot",
          "type": "string",
          "x-go-name": "Error"
        },
        "message": {
          "description": "message",
          "type": "string",
          "x-go-name": "Message"
        },
        "statuscode": {
          "description": "statuscode",
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusCode"
        }
      },
      "x-go-name": "UserTokenBadResponse",
      "x-go-package": "github.com/sparkeh/homer-app/model"
    },
...
  },
  "securityDefinitions": {
    "bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}